% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{check_coverage}
\alias{check_coverage}
\title{Check coverage of panel map over source data}
\usage{
check_coverage(data_in, pm, code_in, code_out, values_from)
}
\description{
Check coverage of panel map over source data
}