% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{check_missing}
\alias{check_missing}
\title{Checks Source Data for Missing Values}
\usage{
check_missing(data_in)
}
\arguments{
\item{data_in}{A Data Frame containing the values you want to transform}
}
\description{
Checks Source Data for Missing Values
}