% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{check_weights}
\alias{check_weights}
\title{Check panel map weights are valid}
\usage{
check_weights(df, code_in, code_out, weights)
}
\arguments{
\item{df}{Data Frame containing weighted links \code{weights} between \code{code_in} and \code{code_out}.}

\item{code_in}{Variable in \code{code_dict} containing source codes to convert from.}

\item{code_out}{Variable in \code{code_dict} containing destination codes to convert to.}

\item{weights}{Column containing weights for transforming values from \code{code_in} to \code{code_out}}
}
\value{
The original data frame if the check is passed and an error if not.
}
\description{
Checks if \code{code_in}, \code{code_out} and \code{weights} columns of data frame forms a valid panel map.
}