% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{concord}
\alias{concord}
\title{Title}
\usage{
concord(.data_in, .map, from_code, to_code, m_weights, ...)
}
\arguments{
\item{.data_in}{A data frame containining \code{from_code} and}

\item{.map}{}

\item{from_code}{}

\item{to_code}{}

\item{...}{variables to transform}
}
\description{
Title
}