% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{conformr-package}
\alias{conformr}
\alias{conformr-package}
\title{Validated Data Harmonisation}
\description{
Provides an opinionated toolkit for harmonising data from related nomenclature or classifications into a single consistent validated dataset. Offers safe-guards against common but often subtle data loss or corruption mistakes.
}
\author{
\strong{Maintainer}: Cynthia Huang \email{cynthia@gmail.com}

Authors:
\itemize{
  \item Laura Puzzello [funder]
}

}
\keyword{internal}