% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_pm_equal}
\alias{make_pm_equal}
\alias{make_panel_map_equal}
\title{Helper to build equal split panel map}
\usage{
make_pm_equal(code_dict, code_in, code_out, .weights_to = NULL)

make_panel_map_equal(code_dict, code_in, code_out, .weights_to = NULL)
}
\arguments{
\item{code_dict}{Data frame containing correspondence between source and destination codes}

\item{code_in}{Variable in \code{code_dict} containing source codes to convert from.}

\item{code_out}{Variable in \code{code_dict} containing destination codes to convert to.}

\item{.weights_to}{(optional) new column name for storing weights that will be applied to. The default name is \verb{split_<<code_in>>}.
input values.}
}
\value{
Returns panel map as tibble
}
\description{
Generate panel map using all \emph{distinct} correspondences between two classifications.
}