% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{use_panel_map}
\alias{use_panel_map}
\title{Apply panel_map to data without checks}
\usage{
use_panel_map(.data, .map, .from, .to, .weights, .vals, .suffix, .by)
}
\arguments{
\item{.data}{a Data Frame assumed to meet Source Data conditions}

\item{.map}{a Data Frame assumed to meet Panel Map conditions}
}
\value{
The output has the following properties:
\itemize{
\item Groups are taken from \code{data_in}
}
}
\description{
A wrapper around a \code{{dplyr}} pipeline that takes a panel_map,
joins it with data, and transforms selected variables in that data according to
instructions in the panel map. Any groups in \code{data_in} are preserved.
}