% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{check_coverage}
\alias{check_coverage}
\title{Check coverage of panel map over source data}
\usage{
check_coverage(data_in, pm, from_code, to_code)
}
\arguments{
\item{data_in}{A Data Frame containing the values you want to transform}

\item{pm}{A Data Frame containing valid Mapping Weights between \code{from_code} and \code{to_code}.}

\item{from_code}{Variable containing Source Codes. Must be present in both \code{data_in} and \code{pm}}

\item{to_code}{Variable in \code{pm} containing Target Codes.}
}
\value{
\code{data_in} if check is successful, throws error otherwise.
}
\description{
Check coverage of panel map over source data
}