% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{concord}
\alias{concord}
\title{Transform data from Source to Target classification using Panel Map}
\usage{
concord(
  data_in,
  pm,
  from_code,
  to_code,
  m_weights,
  values_from,
  .suffix = NULL
)
}
\arguments{
\item{data_in}{A Data Frame containing the values you want to transform}

\item{pm}{A Data Frame containing valid Mapping Weights between \code{from_code} and \code{to_code}.}

\item{from_code}{Variable containing Source Codes. Must be present in both \code{data_in} and \code{pm}}

\item{to_code}{Variable in \code{pm} containing Target Codes.}

\item{m_weights}{Variable in \code{pm} containing Mapping Weights.}

\item{values_from}{A vector of variables in \code{data_in} to be transformed. E.g. \code{c(var1, var2)}}

\item{.suffix}{An (optional) string appended to each \code{values_from} name to create column names for transformed values.
Defaults to \code{"_out"}}
}
\description{
Currently checks for valid Mapping weights, missing values, and coverage.
}
\examples{
\dontrun{
concord(data_in = equal_pm$data_A,
       pm = equal_pm$pm_BA,
       from_code = std_A,
       to_code = std_B,
       m_weights = weight,
       values_from = c(A_100),
       .suffix = "_out")
}

}