% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{has_coverage}
\alias{has_coverage}
\title{Flag if data set is not completely cover by panel map}
\usage{
has_coverage(.data, .map, .from)
}
\arguments{
\item{.data}{a Data Frame assumed to meet Source Data conditions}

\item{.map}{a Data Frame assumed to meet Panel Map conditions}
}
\description{
Flag if data set is not completely cover by panel map
}